Fix tbptt_reduce_fx when non-floating tensors are logged (#3796)
* Add failing test

* force all tbptt vals to be floats for reduce

+++ pytorch_lightning/core/step_result.py
@@ -420,7 +420,7 @@ class Result(Dict):
                 tbptt_reduce_fx = torch.mean
             else:
                 tbptt_reduce_fx = meta[k]['tbptt_reduce_fx']
-            result[k] = tbptt_reduce_fx(value)
+            result[k] = tbptt_reduce_fx(value.float())
 
         result['meta'] = meta
         return result
